WIP: Introduce 'aggregates'
For us to be able to produce information based off an entire MA
we need to be able to manipulate an entire columns worth of data.

Initially, this commit introduces a backend cloned from formulas.js.

Aggregate functions take an array and return a value, included are
three ones to begin with mean, mode, median.

Initial playing with visually displaying these started, there is a new row
in the table beneath 'add a row' however there are questions regarding
how we handle aggregate data, and a new table entirely might be the correct
way to do this.